backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
# Note: python-dotenv is required for test scripts to load .env files
websockets==12.0
aiofiles==23.2.1
Pillow==10.0.1
requests>=2.31.0,<3.0.0
numpy==1.26.2
opencv-python==4.8.1.78
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.3
diffusers==0.24.0
transformers==4.35.2
accelerate==0.25.0
# Video Generation Dependencies
scipy>=1.11.0
psutil==5.9.6
tenacity>=8.2.0

# Post-Processing Dependencies
piexif>=1.1.3  # For metadata removal
realesrgan>=0.3.0  # Real-ESRGAN for upscaling
gfpgan>=1.3.8  # GFPGAN for face restoration

# Optional: Frame interpolation libraries
# rife-ncnn-python  # Uncomment if using RIFE
insightface>=0.7.3  # Face recognition and consistency
onnxruntime>=1.16.0  # Required for InsightFace
imagehash>=4.3.1  # Perceptual hashing for duplicate detection
# basicsr>=1.4.2  # Dependency for realesrgan and gfpgan

# Note: FFmpeg must be installed separately on the system
# Install with: choco install ffmpeg or winget install ffmpeg

# Production Dependencies
redis==5.0.1
prometheus-client==0.19.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Platform Integration Dependencies
instagrapi==2.0.3  # Instagram API
tweepy>=4.14.0  # Twitter/X API
facebook-sdk>=3.1.0  # Facebook Graph API
python-telegram-bot>=20.0  # Telegram Bot API
playwright>=1.40.0  # Browser automation for Instagram, Twitter, OnlyFans
google-api-python-client>=2.100.0  # YouTube Data API
google-auth-httplib2>=0.1.1  # YouTube authentication
google-auth-oauthlib>=1.1.0  # YouTube OAuth
